Week 2/Day 1/requirements.txt
numpy==1.25.2
openai
python-dotenv==1.0.0
pandas
scikit-learn
ipykernel
matplotlib
plotly
arxiv
python-dotenv
faiss-cpu
typing-extensions
tiktoken